Added taskDisplayHook
**Overview**

This is a function the user can set when configuring the plugin. It
will be called on every task to be displayed in the sidebar.

**API**

It should take a task and return a task. The returned task can be
modified, unmodified or `nil`. If `taskDisplayHook` returns `nil`, the
task is skipped.

NOTE: I am not including the `Task` API here nor in the README,
as I don't know how much of it can be considered "public".

**Motivation**

A user may wish to share a single `todo.txt` file between several
clients (using different todo.txt implementations). These clients will
likely have a different set of supported extensions.

It is impossible for a single plugin to implement all of these
extensions. `taskDisplayHook` allows the user to implement their own.

TODO: Whilst this covers modification and filtering, it does not cover
sorting. Perhaps a sort hook (compatible with `table.sort`) can be added
later.